GLOBAL ILLUMINATION!!!!!!!!!
It ain't perfect, and it ain't fast, but it does work. Occlusion texture
needs to be blurred, probably in a shader replacing CPU based coverage
calculation. Lighting info needs to not be locked to the Y axis, values
need to be placed between voxels, and better sampling implemented
alongside performance improvements if possible...